+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Add possibility to define the drop cursor effect
+- Add possibility to get custom items dropped
+### Changed
+- Renamed properties in the drop item
 
 ## [2.0.0] - 2024-06-07
 ### Changed